# [imageio][performance] Add OpenMP to the JPEG2000 loader