This PR significantly speeds up the conversion of decoded image data into the format required by darktable.

For example, on my processor (a rather old Intel i7-6700HQ) on a test JPEG2000 images with a resolution of 24 MP, the following acceleration is observed:

RGB data:
- 0.160 sec without OpenMP
- 0.045 sec with OpenMP

Grayscale data: 
- 0.100 sec without OpenMP
-  0.038 sec with OpenMP

On top of that, the code has been reformatted and the variables have been renamed to make the code more understandable.